ILE: Create files for both ILE classes
This helps clean the code a bit
